Add impl_deps - cc_library implementation deps set
impl_deps provided in cc_library will specify a dependency set which is
only required for the implementation of the library. Implementation-only
dependencies do not contribute to the interface definition of a library,
and are not automatically included transitively for dependents. This is
particularly useful for minimizing the downward propagation of header
inputs to compilations.

This change was later combined with two others. The first being the revert of
the removal of the separation of interface and implementation dependencies and
the second being the addition of two shell tests to verify that implementation
dependencies are working as expected.

The first test verifies that headers in `cc_libraries` included with the
`impl_deps` rule are not visible to `cc_binary` targets later in the chain. The
second test verifies that symbols generated in implementation dependencies are
properly linked and available to `cc_binary` targets later in the chain.