Undo change to introduce methods to remove file system instance from Path.
I no longer believe that it is worthwhile to remove the file system instance from Path. Measurements indicate that I can simply replace the Path implementation in one go and just leave the instance on the Path class.

This CL was done entirely using IDE refactoring, so should be safe.

PiperOrigin-RevId: 180287162