Don't output duplicate srcs when jar and sourcejar are the same
When a source jar is used as maven_jar the filegroup in the generated BUILD.bazel contains duplicate entries. This results in an error when the jar is used as a dependency (... is duplicated in the 'srcs' attribute of rule 'file').

This change simply doesn't write the duplicate entry to the BUILD.bazel file when  the filenames are equal.

Closes #4333.

PiperOrigin-RevId: 180534960