Remove code that used to support --javabase=<absolute path>.
This is the first step in removing package loading from JvmConfigurationLoader (I didn't want to add the rest into this change because it's technically possible to access ctx.fragments.jvm even though it contains no fields, so removing that is an incompatible change) and it's also possible that removing error reporting from JvmConfigurationLoader causes some subtle changes in behavior.

Baby steps. Now that the hard part is done, there is no need to rush.

RELNOTES: None.
PiperOrigin-RevId: 181143978