Make RemoteActionFileSystem not implement MetadataProvider.
The difference this makes in the code is that then the instanceof check in
SkyframeActionExecutor.createFileCache() will not override the
MetadataProvider implementation of ActionMetadataHandler with the action
file system, which is not necessary because all declared input files are also
in ActionInputMap.

The only thing that needed to be fixed is that I had to flip
the order of lookup in the two MetadataProvider instances in
ActionExecutionContext.withOutputsAsInputs() because SingleBuildFileCache
throws an IOException when a file is not found, so the second MetadataProvider
cannot be checked that way.

This confers a slight performance loss (the common case is that the metadata
comes from ActionInputMap), but it's not too bad and it only happens on the
additional Spawns created by StandaloneTestStrategy.

RELNOTES: None.
PiperOrigin-RevId: 523641864
Change-Id: Iebb509eb6a06b688caa2c2f1f2b926cd5c8e6bf2