[Skymeld] Skip organizing ALVs into shards before conflict checking.
For regular blaze, conflict checking lies on the critical path. It's therefore
beneficial to organize the ActionLookupValues into shards to minimize the wall
time there.

With Skymeld, the conflict checking process is done per top level target and
does not necessarily lie on the critical path. The cost of forming the Shards
(essentially Lists of Lists) isn't trivial and it adds up when we do it
multiple times per build, as observed in one of the garbage profiles [1].

This CL skips the organization of ALVs into shards before conflict checking for
Skymeld.

PiperOrigin-RevId: 523676154
Change-Id: Iad433a9998ea0de65220904930c9588463fdee7b