Optimize storage of outputs in `AbstractAction`.
They are currently stored as an `ImmutableSet`. This is wasteful because:

* Most actions only have only a single output.
* Most callers of `getOutputs()` only wish to iterate over the outputs, so optimizing for fast `contains` checks is unnecessary.

Singleton outputs are reconstituted as an `ImmutableSet`. Multiple outputs are reconstituted as a lightweight `Set` wrapper around the backing array. Callers that wish to perform fast lookups can use `ImmutableSet#copyOf`.

I considered adding a garbage-free method to get the number of outputs, but this was not a significant use case.

PiperOrigin-RevId: 523854663
Change-Id: I6e98f8b5f4b99453c3597d79b9b5a14aa244776c