python: port PyInfo to Starlark and enable it.
Because providers are based on identity, and there are Java tests that need
to reference the provider, it's hard to split up the changes, so it's all
in a single change.

A side-effect of this is the name used to access the provider in `cquery` commands
changes slightly: from "PyInfo" to "@_builtins//:common/python/providers.bzl%PyInfo". This fully qualified name is considered internal and shouldn't be relied upon; it will change again when
the rules are removed from Bazel and part of rules_python.

The Java class is kept to make usage by the Java tests easier; it isn't
actually used outside of tests.

Work towards #15897

PiperOrigin-RevId: 524054712
Change-Id: I1ab7d299c7b0458e0ee5359babb1fa63f9739498