Remove the `boolean` field tracking input discovery state from `AbstractAction` in favor of storing it only in action classes which may discover inputs.
As the code comment suggests, this field is a waste for action classes which never discover inputs, e.g. `JavaCompileAction`. It is sometimes also a waste for classes such as `StarlarkAction` which only discovers inputs some of the time, but I want to wait for other refactoring to land before deciding whether it is worthwhile to introduce more subclasses.

For clarity, rename `inputsDiscovered()` to `inputsKnown()`. This makes it more apparent that it should always return true for non-input-discovering actions. The implementation of `inputsKnown()` delegates to a `@ForOverride` method `inputsDiscovered()` for input-discovering actions. Some related Javadoc is consolidated into `ActionExecutionMetadata` where the methods are defined.

There's only a slight memory benefit from this change on its own. That's because the most common action classes still need to store the `boolean` field. However, the unaligned sizes of some classes move closer to an 8-byte threshold due to less waste from padding:

* `SpawnAction`: 80 -> 76
* `CppCompileAction`: 144 -> 140
* `JavaCompileAction`: 96 -> 92

In general, I want to extract out fields that are only needed some of the time. Each individual change may not show much benefit due to padding/alignment, but combining several changes like this will.

PiperOrigin-RevId: 524860596
Change-Id: I1eb46f6644ffa34800db793a8006300930d38183