Narrow inputs for computation of unloaded toolchain contexts.
Instead of requiring a long list of arguments, the tuple composed of
ExecGroupCollection.Builder and the default ToolchainContextKey
suffices.

Part of the logic for auto exec groups has been moved into
ExecGroupCollection where it is more logically cohesive and reduces the
number of map-to-map-to-map transformations by one. Some logic remains in
createDefaultToolchainContextKey.

PiperOrigin-RevId: 524962389
Change-Id: Ic892d17a5433d37252c9bf7e0df1d094a8c6eccc