Store the set of allowed derived inputs in CppCompileAction.
Previously, this was recomputed on every invocation of getAllowedDerivedInputs().

This was quite brittle because the semantics of ArtifactNestedSetKey mean that nested sets are checked for equality by reference most of the time, which in turn means always returning a new NestedSet instance risked redoing computation or even illegal error handling.

RELNOTES: None.
PiperOrigin-RevId: 525454982
Change-Id: Ifbdb3079ae639588d0896b0a4f2914ebb7c4183b