Automated rollback of commit 770c50b286473abef37c1b681f263d93520e9f15.
*** Reason for rollback ***

This CL raises "Exec format error" in some targets.

*** Original change description ***

Move grep-includes to the cc_toolchain

Motivation for this change was an error with the wrong exec platform for grep-includes when Automatic Exec Groups are enabled.

To solve this, grep_includes is moved inside the cc_toolchain, which automatically chooses macos as an exec platform when needed.

PiperOrigin-RevId: 526631776
Change-Id: I48d2d06d41f3d44fd24ef0bc548da90fc79f9361