Fix order of libraries in cc_shared_library
The libraries from different direct dependencies were added to a list without
respecting the topological order. This becomes a problem when a library with
several objects is a dependency of more than one of the direct deps leading
to backref errors.

The CL solves it by properly merging the libraries of different dependencies
using depsets.

RELNOTES:none
PiperOrigin-RevId: 526933703
Change-Id: I91289188a3e4150e6cd97078e838653435e4a846