Make `--experimental_remote_download_regex` work in incremental builds.
When building without the bytes, we can set `--experimental_remote_download_regex` to download additionaly outputs. However, the flag doesn't work with incremental builds, since it depends on action executions. That is, in an incremental build, if you just:

1. deleted previous downloaded outputs
2. or changed the regex

outputs that are expected to be downloaded will not be downloaded since the generating actions are not executed, because of cache.

This change fixes that by checking whether Bazel should download an output during the dirtiness check. If so, we mark generating actions dirty so they will be re-executed (similar to the case in normal build when local outputs are missing).

PiperOrigin-RevId: 527242696
Change-Id: If985d02d1dffd89780885f5b94c171a29986c30f