Add "mandatory spawn inputs" to CppCompileAction.
This is necessary because at HEAD, linkstamp compile actions abuse inputsForInvalidation: linkstamp compile actions need to be re-run if any input of the linker changes, and at HEAD, this is accomplished by putting every linker input in inputsForInvalidation, then they up on the inputs of the action.

The reason why this needs to be changed is that at HEAD, inputsForInvalidation is also the mechanism used for depending on possible inputs discovered by input discovery. These two use cases have different requirements: possible discovered inputs do not need to be part of the action cache key (only discovered ones need to be) but inputs for linkstamp compile action do.

This is accomplished at HEAD by hiding possible discovered inputs behind a scheduling dependency middleman. If one wants to do away with scheduling dependency middlemen, these two use cases need to be distinguished in some other way.

The plan is that possible inputs will be returned by Action.getSchedulingDependencies() and inputs for linkstamp compile invalidation will not be in the inputs of the Spawn returned by CppCompileAction.createSpawn().

Yes, this means that C++ compile actions have a bewildering array of concepts all called "inputs":

- Inputs to the Skyframe action execution function
- Inputs as recorded by the action cache
- Inputs that are sent to the sandbox or remote execution
- Inputs that are actually read by the action

I hope that once --experimental_use_scheduling_middlemen is flipped,this can be simplified a bit because it's very confusing.

Also updated a test so that it doesn't rely on mocking out half of CppCompileAction. Unfortunately, the most convenient new place for it is in a class that hasn't been open sourced yet.

RELNOTES: None.
PiperOrigin-RevId: 527253573
Change-Id: Idcec30b9b50e92f5a1168605f1633be17da83e9b