Remove `PackageValue.Key` wrapper around `PackageIdentifier` in favor of implementing `SkyKey` directly.
This saves a tiny bit of memory but not much since there aren't that many `PACKAGE` nodes. Perhaps more importantly, it saves the cost of constructing a `PackageValue.Key`, which includes interning. This happens even more frequently with pooled `Label` interning, since we need to look up the corresponding `PACKAGE` node in the graph.

PiperOrigin-RevId: 527294657
Change-Id: I297046fbb67c6b9bd956e726bf403caf2f037b09