Allow json.decode to return a default value if decoding fails
It is useful to be able to decode a potentially invalid JSON string without
failing Starlark evaluation. Typically, in this case, we would want to return
a default value - either a reasonable safe default or some domain-specific
representation of "we encountered an error".

Note that the original behavior of json.decode failing on invalid JSON input
still remains (for lack of a better word) the default: the `default` value,
even if it is `None`, must be specified explicitly.

Fixes https://github.com/bazelbuild/bazel/issues/18229

RELNOTES: json.decode now takes an optional `default` positional parameter;
if this parameter is specified and decoding fails, json.decode will return
the `default` value instead of failing Starlark evaluation.
PiperOrigin-RevId: 527958046
Change-Id: I7b0f6698b41024689f1e1a3971b3dbc29647da7a