Add a flag to disallow remote actions from including symlinks in a tree artifact.
A tree artifact is currently defined to contain only regular files. If a remote action reports back a symlink, it must be resolved locally, which is generally impossible and potentially non-hermetic. Allowing only the hermetic cases (i.e., where resolution doesn't escape the action's inputs and outputs) entails additional complexity and is of dubious value. So let's instead ban it through an incompatible change.

Note that this is distinct from #16361, which calls for the ability to include *unresolved* symlinks in a tree artifact on an opt-in basis.

PiperOrigin-RevId: 528731896
Change-Id: Ia48d2ec2fa6d01ce227315c85bbcf5c3f07bb29a