[aquery] Also parallelize the output producer in case of --skyframe_state.
The previous change (https://github.com/bazelbuild/bazel/commit/68702a7affc5bdd6bf924726c39a1cd4bae9de8b) left out this part.

Also fix a concurrency bug with the parallel output. This scenario was possible before:

```
    Main thread: queue.put(POISON_PILL) -> outputStream.flush()
Consumer thread:              <consumer printing stuff> -> queue.take(POISON_PILL)
```

What happened there was the outputStream being flushed (via close()) and the eventual closing of the AqueryOutputHandler while the consumer maybe still be going through the tasks in the queue. This would cause unexpected output from blaze.

By forcing #stopConsumer to wait for the consumer to finish consuming the POISON_PILL, we can avoid this race condition. As POISON_PILL is guaranteed to be the last task, we can be sure that nothing's left to be printed in the output streams by the time we're done with #stopConsumer.

PiperOrigin-RevId: 528773712
Change-Id: Iffc16d27223fdfb7c9d3329f9973593bfb4b2283