Create smart `ActionLookupData` subclasses for action indices 0-9.
This saves the `int` field, reducing the cost of an instance from 24 bytes to 16.

The action index represents the index of the action within the owning target's lists of actions so the most common index is 0, followed by 1, followed by 2, etc. Offer smart subclasses up to a somewhat arbitrary cutoff point.

Unshareable instances are not eligible for the smart subclasses, but these aren't common (mostly test actions and the build info action).

PiperOrigin-RevId: 528800669
Change-Id: I2491023bcc9855a5087859b8405f403749b53d06