Delete .git directory from git_repository external repositories when using strip_prefix
Leaving the `.git` directory in an external repository makes the rule not reproducible since this directory contains timestamps of when the fetch occurred. When strip_prefix is not specified, this directory is correctly cleaned up but not when strip_prefix is specified, since the `.git` directory lives in `.tmp_git_root`.

Note that we cannot delete the whole `.tmp_git_root` directory, since the items in the root are symlinks to here.

Fixes #18152.

Closes #18271.

PiperOrigin-RevId: 528962539
Change-Id: I03b0ee6e6b5464af2d28eb6dc05fb6e6f33afaa3