Change how to download toplevel artifacts.
With https://github.com/bazelbuild/bazel/pull/16524, we download toplevel artifacts by listening to `TargetCompleteEvent` so that it works for incremental builds (e.g. change toplevel targets, delete downloaded files, etc.).

However, BES sends `TargetComplete` event also by listening to `TargetCompleteEvent` which means at the time when the consumers of BEP received `TargetComplete` events, the outputs might not have been downloaded.

This CL changes the way we download toplevel artifacts from listening `TargetCompleteEvent` to checking whether an output is a toplevel artifact after action execution. It still works for incremental builds because of the RemoteOutputChecker we introduced in https://github.com/bazelbuild/bazel/commit/60c9cf3643148b6dce4a6549363f71656dde760c. It doesn't work for BEP yet because the download still happens in the background. A following CL will fix that.

Note that this change makes RemoteOutputChecker depending on *FULL* analysis result in incremental builds because it needs that to determine whether an output is toplevel artifact during action dirtiness check. It's incompatible with skymeld. A way to fix it is described in the code.

PiperOrigin-RevId: 529026614
Change-Id: I6c0c9bf4b3f5bb076f404e46c00bc231789a2820