Disable the most time-consuming tests in presubmit on macOS arm64 platform.
Waiting for an available macOS arm64 machine is often the bottleneck of Bazel presubmit during busy time because we have limited number of machines. This change disabled the top 50 most time-consuming tests in presubmit, but they are still enabled in postsubmit where we can catch breakages after the submitting the change.

PiperOrigin-RevId: 529071647
Change-Id: Ia62a81afcd1053831bf6a60f9ba6872967585e94