[Skymeld] Handle some edge cases with action conflict checking.
For example, when a conflict is caused by an extra action that can't be traced
back to a top level target via TopLevelActionConflictReport.

Also moved the sending of TopLevelTarget/AspectAnalyzedEvent to before action
conflict checking, so that it's consistent with the non-Skymeld code path.

Some integration tests needed to be updated to reflect Skymeld's current known
issue with continuing the build if the conflict turns out to be from unused
actions.

PiperOrigin-RevId: 529405958
Change-Id: Ie06c59dddc29eb1a836a7260ddc0db385461cc1e