Simplify RuleFactory
Previously RuleFactory consisted of static helpers + a simple view wrapping a RuleClassProvider's map from rule name to RuleClass object. This CL eliminates the wrapping behavior so that it's fully static, and moves some rule object logic from PackageFactory over to RuleFactory.

This is part of a larger effort to get PackageFactory out of the business of registering info about the Starlark environment, native rules, and other extensions.

- PackageFactory.buildRuleFunctions() and PackageFactory.BuiltinRuleFunction are moved to RuleFactory.

- PackageFactory.getRuleClass[Names]() was only used by DumpCommand, which is easily changed to use the rule class provider instead.

- The map from rule name to Starlark callable symbol is moved to the rule class provider, and retrieved from there by BazelStarlarkEnvironment.

- DumpCommand and Workspace logic uses the underlying map directly instead of the RuleFactory.

PiperOrigin-RevId: 529877244
Change-Id: I5bdcf56487221c3b0ed176f087b2963d7ce44e4c