Remove _original_source_file from ProtoSourceInfo
There's a small saving in retained heap.

This only uses in experimental_should_generate_code and experimental_filter_sources are replaced with _source_file.

The difference with _original_source_file is that _source_file have import_prefix and strip_import_prefix applied (if prefixes are not set files match).

Technically, the check could be different if .proto file was used by 2 different proto_library rules. However there are not a lot of protos excluded from generation and they are under control by protoc.

PiperOrigin-RevId: 530276398
Change-Id: If0c87d274648015c7822d7135329f346aca47494