Don't set version via PackageFactory
The `version` param to PackageFactory's constructor was only there to be threaded through BazelStarlarkEnvironment to WorkspaceFactory, where it became the value of `native.bazel_version` in WORKSPACE-loaded .bzl files.

In a bazel binary, the value of this param was always retrieved from BlazeVersionInfo. In unit tests and PackageLoader, there were dummy values but they were never checked. (The one test that cares about `native.bazel_version`'s value is an integration test that sees the real value.)

This CL eliminates that param. Instead, WorkspaceFactory retrieves the version from the ConfiguredRuleClassProvider, which always returns the real value.

PiperOrigin-RevId: 530973830
Change-Id: Ia857af0baa62225e2eb7a59ffa78470cd56dbbd0