Remove `ActionEnvironment` field from `AbstractAction`.
Many types of actions always use `ActionEnvironment#EMPTY`. Other actions use `BuildConfigurationValue#getActionEnvironment` and already store the `BuildConfigurationValue`, so don't need the additional field.

For `JavaCompileAction`, which wraps the configuration's environment with an additional entry, create the resulting `CompoundActionEnvironment` on-demand. This is actually where most of the memory savings in this change come from.

`AbstractAction` is now down to just one constructor.

PiperOrigin-RevId: 531200828
Change-Id: Ied9c9c830aef2fef25934eb6e08fe3c75652d6c3