[Skymeld] Update target_compatible_with_test.sh to work with Skymeld too.
With Skymeld, the error would look something like:

```
ERROR: Analysis of target '//foo' failed; build aborted: Target //foo is incompatible and cannot be built, but was explicitly requested.
```

instead of
```
ERROR: Target //foo is incompatible and cannot be built, but was explicitly requested.
```

This CL updates the test to also account for that error message.

PiperOrigin-RevId: 531215789
Change-Id: I508c5cd8166b3b45a1a55e07528fc0721d5b3686