Stop posting a dummy `BuildInfoEvent` if one was already posted.
Currently a dummy event is posted by `BuildTool` at the end of a build to ensure that there's always a `BuildInfoEvent` for every build, including builds that fail fast. However, in the common case, there was already a "real" `BuildInfoEvent`. This oddity shifts responsibility onto subscribers to ignore the second event and has led to confusion on multiple occasions.

There is still a possibility of subscribers noticing the "real" `BuildInfoEvent` twice if it is posted in `BlazeModule#beforeCommand` since `BlazeCommandDispatcher`  replays such events in case subscribers were not yet registered. For this reason, I did not remove subscriber code that deduplicates `BuildInfoEvent`, but at least we no longer post two different events.

PiperOrigin-RevId: 531505934
Change-Id: I4c5d99c8e4a36ed9eb601c06e61d159327946780