Move def of package() out of PackageFactory
The definition of the package() symbol was the last remaining piece of business logic where PackageFactory had a say in what the Starlark environment for BUILD/.bzl files looks like. This CL moves the definition of package() to DefaultPackageArgs (now renamed PackageCallable) and stores the symbol in the rule class provider.

It should now be possible to create a complete Starlark environment for BzlLoadFunction, etc., without knowing the PackageFactory.

PackageCallable now holds both the core package() implementation and the definitions of those package() params that are universal to all Bazel environments. I also simplified by inlining createPackageArguments() into newPackageFunction (renamed newPackageCallable to avoid confusion with, well, PackageFunction).

PiperOrigin-RevId: 531529822
Change-Id: I5f6942717ff439cac85ac612b8e4f92c51c140db