Move import path calculation to _get_import_path
This changes the values of ProtoSource._proto_path and makes them shorter. This is fine, because the field private / not exposed to others. More computation is done in get_import_path.

No _virtual_includes case: Import paths when Bazel repos are involved can be computed from the file.short_path. That is because the path in a repo always starts with '../repo' (even for sibling repository layout). (In native code there's even a function that already strips ../repo).

_virtual_includes case: Only the shorter path is stored (for example a/b/_virtual_includes). get_import_path first removes the repo and then the shorter path.

PiperOrigin-RevId: 532111289
Change-Id: I3d39254bd35c73941a2d67c37fb814b63b2d2762