Optimize the memory footprint of `ArtifactNestedSetKey`.
As explained in the description of https://github.com/bazelbuild/bazel/commit/5bbd8c7edf47f4dd44bdb01d8fa972051516d451, the `NestedSet` and `NestedSet.Node` fields always reference objects that are only referenced from the key. Instead, move `ArtifactNestedSetKey` to the `nestedset` package so that it can get access to the backing `Object[]`. This saves 48 bytes per `ArtifactNestedSetKey`:

* `ArtifactNestedSetKey` reduced from 2 fields to 1: 8 bytes
* `NestedSet.Node`: 16 bytes
* `NestedSet`: 24 bytes

Some of the logic for rewinding through `NestedSet` nodes is moved from `ActionRewindStrategy` to `ArtifactNestedSetKey` because it's easier to implement there without widening access to the backing array.

Additionally, document that `ArtifactNestedSetKey` is based on identity equality, and include the array size in the string representation of `ArtifactNestedSetKey`. This would have been helpful for debugging b/185998331.

PiperOrigin-RevId: 532282856
Change-Id: I37c881dbe03e6678ab34d1b594d112071fa43ab0